drivers: gpu: Limit gpu locking to 350MHz
By default the GPU is heavily throttled once skin temps increase or when the IPA driver locks the A57 clock (on touch key boosting for example). This can result in dropped frames due to the GPU getting locked to minimum clocks even when the device is cold in some cases. This now limits how low the locking can force the GPU clock down to ensure a minimum amount of performance.